pyproject.toml
[project]
name = "real-time-translation"
version = "0.1.0"
description = "Real-time translation application for Zoom meetings"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "deepgram-sdk>=5.3.2",
    "google-genai>=1.58.0",
    "gradio>=6.3.0",
    "langchain>=1.2.6",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=0.2.0",
    "python-dotenv>=1.0.0",
]

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
explicit = true

[tool.uv.sources]
rtms = { index = "testpypi" }

[project.optional-dependencies]
zoom = ["rtms>=0.0.2"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "ruff>=0.8.0"]

[project.scripts]
real-time-translation = "real_time_translation.main:main"
real-time-translation-demo = "real_time_translation.gradio_demo:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/real_time_translation"]

[tool.ruff]
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
